compiler: move interface related code to an object_tree::interfaces module
I need to extract some more helper functions, and the interface
related code is already getting relatively large. Move it to it's own
module to keep object_tree.rs a reasonable size.
